refactor: 패키지 정리 및 방 이름 validate 추가
