+++ src/main/java/chess/domain/game/ChessBoard.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
