+++ src/main/java/chess/domain/game/MoveChecker.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
