+++ src/main/java/chess/domain/game/Status.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 public class Status {
