+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -4,6 +4,8 @@ import static chess.constants.TestConstants.PARAMETERIZED_TEST_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.domain.game.ChessBoard;
+import chess.domain.game.GameCommand;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
