+++ src/test/java/chess/domain/ChessRoomTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import chess.domain.game.ChessRoom;
 import org.junit.jupiter.api.Test;
 
 class ChessRoomTest {
