+++ src/test/java/chess/domain/CommandTypeTest.java
@@ -3,6 +3,7 @@ package chess.domain;
 import static chess.constants.TestConstants.PARAMETERIZED_TEST_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.game.CommandType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
