+++ src/test/java/chess/domain/ScoreCalculatorTest.java
@@ -2,6 +2,7 @@ package chess.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import chess.domain.game.ScoreCalculator;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
 import chess.domain.piece.Knight;
